@@ -1,41 +1,30 @@
 'use strict';
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import Site from './../site/site.jsx';
-import Docs from './../docs/docs.jsx';
 import Radio from './../radio/radio.jsx';
 import './scrollbar.css';
 
-export default class Router extends React.Component {
-    constructor() {
-        super();
-        this.state = {
-            page: ''
-        };
-    }
+const Router = () => {
+    const [page, setPage] = useState('');
+    const [loaded, setLoaded] = useState(false);
 
-    componentDidMount() {
-        const url = window.location.href.split('/');
-        const page = url[3] === '' || url[3] === '//' ? '/' : url[3];
-        const pageArg = url[4] ? url[4] : '';
-        if (page && page.substr(0, 4) === 'docs') {
-            this.setState({
-                page: <Docs command={pageArg} />
-            });
+    useEffect(() => {
+        if (!loaded) {
+            const url = window.location.href.split('/');
+            const wLocation = url[3] === '' || url[3] === '//' ? '/' : url[3];
+            const pageArg = url[4] ? url[4] : '';
+            setPage(<Site page={wLocation} urlArg={pageArg} />);
+            setLoaded(true);
         }
-        else {
-            this.setState({
-                page: <Site page={page} urlArg={pageArg} />
-            });
-        }
-    }
+    });
 
-    render() {
-        return (
-            <span>
-                <Radio />
-                {this.state.page}
-            </span>
-        );
-    }
-}
+    return (
+        <span>
+            <Radio />
+            {page}
+        </span>
+    );
+};
+
+export default Router;
