@@ -2,12 +2,12 @@
 
 import React from 'react';
 
-export default class Craft extends React.Component {
-    render() {
-        return (
-            <div className='principal-container'>
-                <h1 className='craft-title text-center col-12 col-md-6'>Crafts</h1>
-            </div>
-        );
-    }
-}
+const Craft = () => {
+    return (
+        <div className='principal-container'>
+            <h1 className='craft-title text-center col-12 col-md-6'>Crafts</h1>
+        </div>
+    );
+};
+
+export default Craft;
