@@ -1,105 +1,89 @@
 'use strict';
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import Axios from 'axios';
 import _ from 'lodash';
-import dragoJSON from './../../../../../assets/json/dragodindes.json';
+import dragoJSON from '~/assets/json/dragodindes.json';
 import DragodindesModal from './modal.jsx';
 import { Tooltip } from '@material-ui/core';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { library } from '@fortawesome/fontawesome-svg-core';
-import { faHeart, faHeartBroken, faToggleOff, faToggleOn } from '@fortawesome/free-solid-svg-icons';
+import { faHeart, faHeartBroken, faToggleOff, faToggleOn, faCheck, faTimesCircle } from '@fortawesome/free-solid-svg-icons';
 import PropTypes from 'prop-types';
 
 library.add(faHeart);
 library.add(faHeartBroken);
 library.add(faToggleOff);
 library.add(faToggleOn);
+library.add(faCheck);
+library.add(faTimesCircle);
 
-export default class Dragodindes extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            user: this.props.user,
-            dragodindes: [],
-            showedDragodindes: [],
-            selectedDrago: [],
-            wait: true,
-            show: false,
-            action: 'default',
-            title: '',
-            input: false,
-            dragoJSON: dragoJSON.sort((a, b) => a.name.localeCompare(b.name))
-        };
-        this.handleAddModalDrago = this.handleAddModalDrago.bind(this);
-        this.handleCallAPI = this.handleCallAPI.bind(this);
-        this.handleClose = this.handleClose.bind(this);
-        this.handleChange = this.handleChange.bind(this);
-        this.handleChangeModal = this.handleChangeModal.bind(this);
-    }
+const Dragodindes = (props) => {
+    const [user] = useState(props.user);
+    const [dragodindes, setDragodindes] = useState([]);
+    const [showedDragodindes, setShowedDragodindes] = useState([]);
+    const [selectedDrago, setSelectedDrago] = useState([]);
+    const [wait, setWait] = useState(true);
+    const [loaded, setLoaded] = useState(false);
+    const [show, setShow] = useState(false);
+    const [action, setAction] = useState('default');
+    const [title, setTitle] = useState('');
+    const [input, setInput] = useState(false);
+    const [dragoJSONConst, setDragoJSONConst] = useState(dragoJSON.sort((a, b) => a.name.localeCompare(b.name)));
 
-    componentDidMount() {
-        Axios.post('/api/dofus/dragodindes', {
-            userId: this.state.user.id
-        })
-            .then(res => {
-                if (res.data && res.data.length) {
-                    this.setState({
-                        dragodindes: res.data.sort((a, b) => a.name.localeCompare(b.name)),
-                        showedDragodindes: res.data.sort((a, b) => a.name.localeCompare(b.name)),
-                        wait: false
-                    });
-                }
-                else {
-                    this.setState({
-                        wait: false
-                    });
-                }
-            });
-    }
-
-    updateDragodindesAPI(url, paramsObj) {
-        Axios.post(url, paramsObj)
-            .then(res => {
-                this.setState({
-                    dragodindes: res.data.sort((a, b) => a.name.localeCompare(b.name)),
-                    showedDragodindes: res.data.sort((a, b) => a.name.localeCompare(b.name)),
-                    action: 'default'
+    useEffect(() => {
+        if (wait && !loaded) {
+            const getDragodindesAPI = async () => {
+                const res = await Axios.post('/api/dofus/dragodindes', {
+                    userId: user.id
                 });
-            });
-    }
+                if (res.data && res.data.length) {
+                    setDragodindes(res.data.sort((a, b) => a.name.localeCompare(b.name)));
+                    setShowedDragodindes(res.data.sort((a, b) => a.name.localeCompare(b.name)));
+                }
+                setWait(false);
+            };
+            getDragodindesAPI();
+            setLoaded(true);
+        }
+    });
 
-    handleAddModalDrago(name, duration, generation, selected) {
-        let dragodindes = this.state.selectedDrago;
+    const updateDragodindesAPI = async (url, paramsObj) => {
+        const res = await Axios.post(url, paramsObj);
+        setDragodindes(res.data.sort((a, b) => a.name.localeCompare(b.name)));
+        setShowedDragodindes(res.data.sort((a, b) => a.name.localeCompare(b.name)));
+        setAction('default');
+    };
+
+    const handleAddModalDrago = (name, duration, generation, selected) => {
+        let localDragodindes = selectedDrago;
         let JSONselected = '';
         if (!selected) {
-            JSONselected = this.setDragoJSON(dragoJSON, name);
-            dragodindes.push({ name: name, duration: duration, generation: generation });
+            JSONselected = setDragoJSON(dragoJSON, name);
+            localDragodindes.push({ name: name, duration: duration, generation: generation });
         }
         else {
-            dragodindes = this.removeDragodindeFromArray(name, dragodindes);
-            JSONselected = this.setRemoveDragoJSON(name);
+            localDragodindes = removeDragodindeFromArray(name, localDragodindes);
+            JSONselected = setRemoveDragoJSON(name);
         }
-        this.setState({
-            selectedDrago: dragodindes.sort((a, b) => a.name.localeCompare(b.name)),
-            dragoJSON: this.state.input ? JSONselected.filter(d => d.name.toLowerCase().indexOf(this.state.input) !== -1).sort((a, b) => a.name.localeCompare(b.name)) : JSONselected.sort((a, b) => a.name.localeCompare(b.name))
-        });
-    }
+        setSelectedDrago(localDragodindes.sort((a, b) => a.name.localeCompare(b.name)));
+        setDragoJSONConst(input ? JSONselected.filter(d => d.name.toLowerCase().indexOf(input) !== -1).sort((a, b) => a.name.localeCompare(b.name)) : JSONselected.sort((a, b) => a.name.localeCompare(b.name)));
+    };
 
-    removeDragodindeFromArray(name, dragodindes) {
-        const array = dragodindes;
-        dragodindes.map((drago, index) => {
+    const removeDragodindeFromArray = (name, localDragodindes) => {
+        const arrayDrago = localDragodindes;
+        localDragodindes.map((drago, index) => {
             if (drago.name === name) {
-                delete array[index];
+                delete arrayDrago[index];
             }
         });
-        return _.compact(array);
-    }
+        return _.compact(arrayDrago);
+    };
 
-    setRemoveDragoJSON(name) {
+    const setRemoveDragoJSON = (name) => {
         const JSONselected = [];
         dragoJSON.map(d => {
-            const check = this.checkAlreadySelected(d);
+            const check = checkAlreadySelected(d);
             if (d.name === name || !check) {
                 JSONselected.push(d);
             }
@@ -108,12 +92,12 @@ export default class Dragodindes extends React.Component {
             }
         });
         return JSONselected;
-    }
+    };
 
-    setDragoJSON(array, name) {
+    const setDragoJSON = (array, name) => {
         const JSONselected = [];
         array.map(d => {
-            const check = this.checkAlreadySelected(d);
+            const check = checkAlreadySelected(d);
             if (d.name === name || check) {
                 JSONselected.push({ ...d, selected: true });
             }
@@ -122,109 +106,97 @@ export default class Dragodindes extends React.Component {
             }
         });
         return JSONselected;
-    }
+    };
 
-    checkAlreadySelected(dragodinde) {
+    const checkAlreadySelected = (dragodinde) => {
         let check = false;
-        this.state.selectedDrago.map(drago => {
+        selectedDrago.map(drago => {
             if (drago.name === dragodinde.name) {
                 check = true;
             }
         });
         return check;
-    }
+    };
 
-    handleChange(e) {
-        const filtered = this.state.dragodindes.filter(d => d.name.toLowerCase().indexOf(e.target.value.toLowerCase()) !== -1);
-        this.setState({
-            showedDragodindes: filtered.sort((a, b) => a.name.localeCompare(b.name))
-        });
-    }
+    const handleChange = (e) => {
+        const filtered = dragodindes.filter(d => d.name.toLowerCase().indexOf(e.target.value.toLowerCase()) !== -1);
+        setShowedDragodindes(filtered.sort((a, b) => a.name.localeCompare(b.name)));
+    };
 
-    handleChangeModal(e) {
+    const handleChangeModal = (e) => {
         const filtered = dragoJSON.filter(d => d.name.toLowerCase().indexOf(e.target.value.toLowerCase()) !== -1);
-        this.setState({
-            dragoJSON: this.setDragoJSON(filtered, false).sort((a, b) => a.name.localeCompare(b.name)),
-            input: e.target.value === '' ? false : e.target.value.toLowerCase()
-        });
-    }
+        setDragoJSONConst(setDragoJSON(filtered, false).sort((a, b) => a.name.localeCompare(b.name)));
+        setInput(e.target.value === '' ? false : e.target.value.toLowerCase());
+    };
 
-    handleCallAPI(url) {
-        if (this.state.selectedDrago.length) {
-            this.updateDragodindesAPI(url, {
-                userId: this.state.user.id,
-                dragodindes: this.state.selectedDrago
+    const handleCallAPI = (url) => {
+        if (selectedDrago.length) {
+            updateDragodindesAPI(url, {
+                userId: user.id,
+                dragodindes: selectedDrago
             });
-            this.handleClose();
+            handleClose();
         }
-    }
+    };
 
-    handleSelectMyDragodindes(name, duration, generation, selected) {
-        const dragodindes = this.state.showedDragodindes;
-        let selectedDrago = this.state.selectedDrago;
-        dragodindes.map(drago => {
+    const handleSelectMyDragodindes = (name, duration, generation, selected) => {
+        const _showedDragodindes = showedDragodindes;
+        let _selectedDrago = selectedDrago;
+        _showedDragodindes.map(drago => {
             if (drago.name === name) {
                 drago.selected = !selected;
             }
         });
         if (selected) {
-            this.state.selectedDrago.map((drago, index) => {
+            selectedDrago.map((drago, index) => {
                 if (drago.name === name) {
-                    delete selectedDrago[index];
-                    selectedDrago = _.compact(selectedDrago);
+                    delete _selectedDrago[index];
+                    _selectedDrago = _.compact(_selectedDrago);
                 }
             });
         }
         else {
-            selectedDrago.push({ name: name, duration: duration, generation: generation });
+            _selectedDrago.push({ name: name, duration: duration, generation: generation });
+            _selectedDrago = _.compact(_selectedDrago);
         }
-        const action = selectedDrago.length === 0 ? 'default' : 'remove';
-        this.setState({
-            showedDragodindes: dragodindes.sort((a, b) => a.name.localeCompare(b.name)),
-            selectedDrago: selectedDrago,
-            action: action
-        });
-    }
+        setSelectedDrago(_selectedDrago);
+        setShowedDragodindes(_showedDragodindes.sort((a, b) => a.name.localeCompare(b.name)));
+        setAction(_selectedDrago.length === 0 ? 'default' : 'remove');
+    };
 
-    handleUnusedDragodindes(drago) {
-        const selected = this.state.selectedDrago;
+    const handleUnusedDragodindes = (drago) => {
+        const selected = selectedDrago;
         selected.push({ name: drago.name });
-        this.setState({
-            selectedDrago: selected
-        });
+        setSelectedDrago(selected);
         if (!drago.used) {
-            this.showModal('unused');
+            showModal('unused');
         }
         else {
-            this.showModal('remove-unused');
+            showModal('remove-unused');
         }
-    }
+    };
 
-    handleLastDragodinde(drago) {
-        const selected = this.state.selectedDrago;
+    const handleLastDragodinde = (drago) => {
+        const selected = selectedDrago;
         selected.push({ name: drago.name });
-        this.setState({
-            selectedDrago: selected
-        });
+        setSelectedDrago(selected);
         if (!drago.last.status) {
-            this.showModal('last');
+            showModal('last');
         }
         else {
-            this.showModal('remove-last');
+            showModal('remove-last');
         }
-    }
+    };
 
-    handleClose() {
-        this.setState({
-            show: false,
-            title: '',
-            selectedDrago: [],
-            dragoJSON: dragoJSON.sort((a, b) => a.name.localeCompare(b.name)),
-            input: false
-        });
-    }
+    const handleClose = () => {
+        setShow(false);
+        setTitle('');
+        setSelectedDrago([]);
+        setDragoJSONConst(dragoJSON.sort((a, b) => a.name.localeCompare(b.name)));
+        setInput(false);
+    };
 
-    showModal(choice) {
+    const showModal = (choice) => {
         let goodTitle = '';
         if (choice === 'new') {
             goodTitle = 'Ajouter des dragodindes';
@@ -241,157 +213,153 @@ export default class Dragodindes extends React.Component {
         else if (choice === 'remove-unused') {
             goodTitle = 'Définir comme disponible ?';
         }
-        this.setState({
-            show: true,
-            title: goodTitle
-        });
-    }
+        setShow(true);
+        setTitle(goodTitle);
+    };
 
-    render() {
-        return (
-            <div className='principal-container'>
-                <DragodindesModal
-                    handleClose={this.handleClose}
-                    handleChangeModal={this.handleChangeModal}
-                    handleAddModalDrago={this.handleAddModalDrago}
-                    handleCallAPI={this.handleCallAPI}
-                    show={this.state.show}
-                    title={this.state.title}
-                    dragodindes={this.state.dragodindes}
-                    dragoJSON={this.state.dragoJSON}
-                />
-                <h2 className='craft-title text-center'>Mes dragodindes</h2>
-                <div className='notes-btn col-sm-12 text-center'>
-                    {!this.state.selectedDrago.length ?
-                        <button
-                            className='my-dragodindes-add-btn'
-                            onClick={() => this.showModal('new')}
-                        >
-                            Ajouter des dragodindes
-                        </button> :
-                        <button
-                            className='my-dragodindes-add-btn'
-                            disabled
-                        >
-                            Ajouter des dragodindes
-                        </button>}
-                    {this.state.action === 'remove' ?
-                        <button
-                            className='my-dragodindes-add-btn'
-                            onClick={() => this.handleCallAPI('/api/dofus/dragodindes/remove')}
-                        >
-                            Supprimer la sélection
-                        </button> :
-                        <button
-                            className='my-dragodindes-add-btn'
-                            disabled
-                        >
-                            Supprimer la sélection
-                        </button>}
-                    <a href='/dofus/fecondator'>
-                        <button>Fécondator</button>
-                    </a>
-                </div>
-                {this.state.dragodindes.length ?
-                    <div className='text-center principal-dragodindes-div col-sm-11 col-md-10 col-lg-8 col-xl-6'>
-                        <input
-                            className='input-parcho'
-                            placeholder='Rechercher'
-                            onChange={this.handleChange}
-                        />
-                        <div className='my-dragodindes-container'>
-                            {this.state.showedDragodindes.length ?
-                                this.state.showedDragodindes.map((drago, index) => {
-                                    return (
-                                        <div
-                                            className={drago.selected ? 'my-drago-line-selected' : drago.last.status ? 'my-drago-line-last' : drago.used ? 'my-drago-line-used' : 'my-drago-line'}
-                                            key={index}
-                                        >
-                                            <div className='my-dragodindes-name col-9'>
-                                                <img src={'/assets/img/dragodindes/' + drago.name.toLowerCase().split(' ').join('-') + '.png'} alt='dd_icon' />
-                                                {drago.last.status ?
-                                                    <p>{drago.name}<span className='my-drago-fecond-message'> - Fécondée</span></p> : drago.used ?
-                                                        <p>{drago.name}<span className='my-drago-used-message'> - Utilisée</span></p> :
-                                                        <p>{drago.name}</p>}
-                                            </div>
-                                            <div className='my-dragodindes-icons col-3'>
-                                                {!drago.used && this.state.action !== 'remove' ?
+    return (
+        <div className='principal-container'>
+            <DragodindesModal
+                handleClose={handleClose}
+                handleChangeModal={handleChangeModal}
+                handleAddModalDrago={handleAddModalDrago}
+                handleCallAPI={handleCallAPI}
+                show={show}
+                title={title}
+                dragodindes={dragodindes}
+                dragoJSON={dragoJSONConst}
+            />
+            <h2 className='craft-title text-center'>Mes dragodindes</h2>
+            <div className='notes-btn col-sm-12 text-center'>
+                {!selectedDrago.length ?
+                    <button
+                        className='my-dragodindes-add-btn'
+                        onClick={() => showModal('new')}
+                    >
+                        Ajouter des dragodindes
+                    </button> :
+                    <button
+                        className='my-dragodindes-add-btn'
+                        disabled
+                    >
+                        Ajouter des dragodindes
+                    </button>}
+                {action === 'remove' ?
+                    <button
+                        className='my-dragodindes-add-btn'
+                        onClick={() => handleCallAPI('/api/dofus/dragodindes/remove')}
+                    >
+                        Supprimer la sélection
+                    </button> :
+                    <button
+                        className='my-dragodindes-add-btn'
+                        disabled
+                    >
+                        Supprimer la sélection
+                    </button>}
+                <a href='/dofus/fecondator'>
+                    <button>Fécondator</button>
+                </a>
+            </div>
+            {dragodindes.length ?
+                <div className='text-center principal-dragodindes-div col-sm-11 col-md-10 col-lg-8 col-xl-6'>
+                    <input
+                        className='input-parcho'
+                        placeholder='Rechercher'
+                        onChange={handleChange}
+                    />
+                    <div className='my-dragodindes-container'>
+                        {showedDragodindes.length ?
+                            showedDragodindes.map((drago, index) => {
+                                return (
+                                    <div
+                                        className={drago.selected ? 'my-drago-line-selected' : drago.last.status ? 'my-drago-line-last' : drago.used ? 'my-drago-line-used' : 'my-drago-line'}
+                                        key={index}
+                                    >
+                                        <div className='my-dragodindes-name col-9'>
+                                            <img src={'/assets/img/dragodindes/' + drago.name.toLowerCase().split(' ').join('-') + '.png'} alt='dd_icon' />
+                                            {drago.last.status ?
+                                                <p>{drago.name}<span className='my-drago-fecond-message'> - Fécondée</span></p> : drago.used ?
+                                                    <p>{drago.name}<span className='my-drago-used-message'> - Utilisée</span></p> :
+                                                    <p>{drago.name}</p>}
+                                        </div>
+                                        <div className='my-dragodindes-icons col-3'>
+                                            {!drago.used && action !== 'remove' ?
+                                                <Tooltip
+                                                    title='Définir comme déjà utilisée'
+                                                    placement='top'
+                                                >
+                                                    <span
+                                                        className='icon-used'
+                                                        onClick={() => handleUnusedDragodindes(drago)}
+                                                    >
+                                                        <FontAwesomeIcon icon='toggle-off' />
+                                                    </span>
+                                                </Tooltip> : drago.used && action !== 'remove' ?
                                                     <Tooltip
-                                                        title='Définir comme déjà utilisée'
+                                                        title='Définir comme disponible'
                                                         placement='top'
                                                     >
                                                         <span
                                                             className='icon-used'
-                                                            onClick={() => this.handleUnusedDragodindes(drago)}
+                                                            onClick={() => handleUnusedDragodindes(drago)}
                                                         >
-                                                            <FontAwesomeIcon icon='toggle-off' />
+                                                            <FontAwesomeIcon icon='toggle-on' />
                                                         </span>
-                                                    </Tooltip> : drago.used && this.state.action !== 'remove' ?
-                                                        <Tooltip
-                                                            title='Définir comme disponible'
-                                                            placement='top'
-                                                        >
-                                                            <span
-                                                                className='icon-used'
-                                                                onClick={() => this.handleUnusedDragodindes(drago)}
-                                                            >
-                                                                <FontAwesomeIcon icon='toggle-on' />
-                                                            </span>
-                                                        </Tooltip> : ''}
-                                                {(!this.state.selectedDrago.length || this.state.show) && !drago.last.status ?
+                                                    </Tooltip> : ''}
+                                            {(!selectedDrago.length || show) && !drago.last.status ?
+                                                <Tooltip
+                                                    title='Définir comme la dernière dragodinde fécondée'
+                                                    placement='top'
+                                                >
+                                                    <span
+                                                        className='icon-heart'
+                                                        onClick={() => handleLastDragodinde(drago)}
+                                                    >
+                                                        <FontAwesomeIcon icon='heart' />
+                                                    </span>
+                                                </Tooltip> : (!selectedDrago.length || show) && drago.last.status ?
                                                     <Tooltip
-                                                        title='Définir comme la dernière dragodinde fécondée'
+                                                        title='Retirer la fécondation'
                                                         placement='top'
                                                     >
                                                         <span
                                                             className='icon-heart'
-                                                            onClick={() => this.handleLastDragodinde(drago)}
+                                                            onClick={() => handleLastDragodinde(drago)}
                                                         >
-                                                            <FontAwesomeIcon icon='heart' />
-                                                        </span>
-                                                    </Tooltip> : (!this.state.selectedDrago.length || this.state.show) && drago.last.status ?
-                                                        <Tooltip
-                                                            title='Retirer la fécondation'
-                                                            placement='top'
-                                                        >
-                                                            <span
-                                                                className='icon-heart'
-                                                                onClick={() => this.handleLastDragodinde(drago)}
-                                                            >
-                                                                <FontAwesomeIcon icon='heart-broken' />
-                                                            </span>
-                                                        </Tooltip> : ''}
-                                                {this.state.action !== 'unused' ?
-                                                    <Tooltip title={drago.selected ? 'Retirer de la sélection' : 'Sélectionner pour la suppression'} placement='top'>
-                                                        <span
-                                                            className='icon-cross'
-                                                            onClick={() => this.handleSelectMyDragodindes(drago.name, drago.duration, drago.generation, drago.selected || false)}
-                                                        >
-                                                            <FontAwesomeIcon icon='times-circle' />
+                                                            <FontAwesomeIcon icon='heart-broken' />
                                                         </span>
                                                     </Tooltip> : ''}
-                                            </div>
+                                            {action !== 'unused' ?
+                                                <Tooltip title={drago.selected ? 'Retirer de la sélection' : 'Sélectionner pour la suppression'} placement='top'>
+                                                    <span
+                                                        className='icon-cross'
+                                                        onClick={() => handleSelectMyDragodindes(drago.name, drago.duration, drago.generation, drago.selected)}
+                                                    >
+                                                        <FontAwesomeIcon icon='times-circle' />
+                                                    </span>
+                                                </Tooltip> : ''}
                                         </div>
-                                    );
-                                }) :
-                                <div className='empty-drago-line'>
-                                    <div className='my-dragodindes-name'>
-                                        <p>Aucun résultat</p>
                                     </div>
-                                </div>}
-                        </div>
-                    </div> :
-                    this.state.wait ?
-                        <div className='text-center loading-notes-message'>
-                            <h1>Chargement des dragodindes <span className='custom-spinner-notes' /></h1>
-                        </div> :
+                                );
+                            }) :
+                            <div className='empty-drago-line'>
+                                <div className='my-dragodindes-name'>
+                                    <p>Aucun résultat</p>
+                                </div>
+                            </div>}
+                    </div>
+                </div> :
+                wait ?
+                    <div className='text-center loading-notes-message'>
+                        <h1>Chargement des dragodindes <span className='custom-spinner-notes' /></h1>
+                    </div> : !wait && loaded ?
                         <div className='text-center no-notes-message'>
                             <h1>Pas de dragodindes actuellement</h1>
-                        </div>}
-            </div>
-        );
-    }
-}
+                        </div> : <></>}
+        </div>
+    );
+};
 
 Dragodindes.propTypes = {
     user: PropTypes.oneOfType([
@@ -399,3 +367,5 @@ Dragodindes.propTypes = {
         PropTypes.bool.isRequired
     ])
 };
+
+export default Dragodindes;
