@@ -2,10 +2,10 @@
 
 import React from 'react';
 
-export default class Parchment extends React.Component {
-    render() {
-        return (
-            <></>
-        );
-    }
-}
+const Parchment = () => {
+    return (
+        <></>
+    );
+};
+
+export default Parchment;
